requirements_service.txt
# Instagram Service - 클라이언트 프로그램 의존성
setuptools>=70.0
selenium==4.15.2
pandas==2.1.3
openpyxl==3.1.2
undetected-chromedriver==3.5.3
chromedriver-autoinstaller==0.6.4
sqlalchemy==2.0.23
nicegui>=2.5.0
pyyaml==6.0.2
cryptography==45.0.2
requests==2.32.3
apscheduler==3.6.3
